+++ platformio/builder/tools/platformio.py
@@ -166,9 +166,7 @@ def ProcessProjectDeps(env):
         projenv.BuildSources(
             "$BUILD_TEST_DIR", "$PROJECT_TEST_DIR", "$PIOTEST_SRC_FILTER"
         )
-    if "test" not in env.GetBuildType() or env.GetProjectOption(
-        "test_build_project_src"
-    ):
+    if "test" not in env.GetBuildType() or env.GetProjectOption("test_build_src"):
         projenv.BuildSources(
             "$BUILD_SRC_DIR", "$PROJECT_SRC_DIR", env.get("SRC_FILTER")
         )